+++ vprdb/io/__init__.py
@@ -1,2 +1,2 @@
-from vprdb.io.exporter import export
-from vprdb.io.reader import read_dataset_from_depth, read_dataset_from_point_clouds
+from vprdb.io.export_utils import export
+from vprdb.io.read_utils import read_dataset_from_depth, read_dataset_from_point_clouds
